requirements.txt
streamlit
openai 
langchain
langchain_openai
langchain_pinecone
langchain_community
langchain[docarray]
docarray
pydantic==1.10.8
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
git+https://github.com/openai/whisper.git